Extract testing code for kalman filter into templated functions
